requirements.txt
# ============================================
# Requirements - Analyses NLP Text Mining
# ============================================

# Base
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# NLP Processing
nltk>=3.8.0
spacy>=3.7.0
scikit-learn>=1.3.0
gensim>=4.3.0

# Embeddings & Deep Learning
sentence-transformers>=2.2.0
torch>=2.0.0

# Dimensionality Reduction & Clustering
umap-learn>=0.5.4
hdbscan>=0.8.33

# Visualizations
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
wordcloud>=1.9.0
pyLDAvis>=3.4.0

# Maps
folium>=0.14.0
geopy>=2.4.0

# Database
duckdb>=0.9.0
psycopg2-binary>=2.9.9

# Web App (pour plus tard)
streamlit>=1.28.0

# Utils
tqdm>=4.66.0
joblib>=1.3.0

# Optional (performance)
numba>=0.58.0

# NLP
spacy>=3.7.0
nltk>=3.8.0
sentence-transformers>=2.2.0

# Network
networkx>=3.2.0
pyvis>=0.3.2

# Utils
tqdm>=4.66.0
openpyxl>=3.1.0

# PostgreSQL (Supabase cloud) 
psycopg2-binary>=2.9.9

python-dotenv>=1.0.0
bs4
sentence-transformers>=2.2.0
selenium
mistralai
